+++ read-matrix-market.hpp
@@ -1,6 +1,6 @@
 #include <vector>
 
-typedef std::vector<std::vector<int>> Matrix;
+typedef std::vector<std::vector<double>> Matrix;
 
 Matrix read_matrix_market(const char* filename);
